feat/generic
Move cmd error into its own module
